requirements.txt
# Multi-LLM Group Chat App - Requirements
streamlit>=1.30.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
